Fixed padding of progress and error indicators on messages
