+++ lib/messaging/widgets/attachment_types/audio.dart
@@ -1,9 +1,10 @@
 import 'dart:typed_data';
 
-import 'package:lantern/messaging/widgets/attachment_builder.dart';
+import 'package:lantern/messaging/widgets/attachment.dart';
 import 'package:lantern/messaging/widgets/voice_recorder/audio_widget.dart';
 import 'package:lantern/model/protos_flutteronly/messaging.pb.dart';
 import 'package:lantern/package_store.dart';
+import 'package:sizer/sizer.dart';
 
 /// An attachment that shows an audio player.
 class AudioAttachment extends StatelessWidget {
@@ -17,6 +18,7 @@ class AudioAttachment extends StatelessWidget {
     return AttachmentBuilder(
         attachment: attachment,
         inbound: inbound,
+        padAttachment: false,
         defaultIcon: Icons.image,
         builder: (BuildContext context, Uint8List thumbnail) {
           return AudioWidget(
@@ -25,7 +27,7 @@ class AudioAttachment extends StatelessWidget {
             initialColor: inbound ? Colors.black : Colors.white,
             progressColor: inbound ? outboundMsgColor : inboundMsgColor,
             backgroundColor: inbound ? inboundBgColor : outboundBgColor,
-            width: MediaQuery.of(context).size.width * 0.5,
+            width: 50.w,
             waveHeight: 50,
           );
         });
