+++ lib/messaging/widgets/attachment_types/image.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:lantern/messaging/widgets/attachment_builder.dart';
+import 'package:lantern/messaging/widgets/attachment.dart';
 import 'package:lantern/model/protos_flutteronly/messaging.pb.dart';
 import 'package:lantern/package_store.dart';
 import 'package:lantern/ui/widgets/basic_memory_image.dart';
